Add more thorough EK policy test, clean up session nonce interface
This change adds a combinatoric EK policy test that tests every
conceivable combination of options for using the EK policy session
(e.g., for decryption, in a salted/bound session, etc) as well
as using both the just-in-time `Policy` and standalone `PolicySession`
session constructors.

This change also fixes the interface of tpm2.Session so that
NonceTPM() returns a TPM2BNonce, for more natural use in commands that
want nonces.