Use Int.FillBytes to left-pad the secret point.
elliptic.CurveParams.ScalarMult returns the point as a pair of
big.Integers. Calling `.Bytes()` on these will return the compacted
representation of the number. In the case of the TPM ECDH secret
agreement, the Z-point's X-value must be left-padded with zeroes to be
the same length as the curve's bit size.

Before the fix, the function would produce an incorrect value around
1/256 of the time (when the leading byte is 0x00). Tested by running the
tests (which perform dozens of such calculations) 100x.